Update storage/level tests
Was getting error "failed to fill whole buffer" from test:
`storage::level::test::create_level`. I think it was just an
issue caused in the test and not in the implementation itself
but I'll keep an eye out.

I *did* get that error before my last commit but it was only
happening intermittently so that could still be the case.